Fix GuardZoneSetting/RadarControl jumping on Linux.
Be very explicit about where we want the control. Do not allow overrides by OS default positioning.

+++ src/br24radar_pi.h
@@ -406,11 +406,11 @@ private:
 
     SOCKET                    m_radar_socket;
 
-    long                      m_BR24Controls_dialog_sx, m_BR24Controls_dialog_sy ;
-    long                      m_BR24Controls_dialog_x, m_BR24Controls_dialog_y ;
+    int                       m_BR24Controls_dialog_sx, m_BR24Controls_dialog_sy ;
+    int                       m_BR24Controls_dialog_x, m_BR24Controls_dialog_y ;
 
-    long                      m_Guard_dialog_sx, m_Guard_dialog_sy ;
-    long                      m_Guard_dialog_x, m_Guard_dialog_y ;
+    int                       m_Guard_dialog_sx, m_Guard_dialog_sy ;
+    int                       m_Guard_dialog_x, m_Guard_dialog_y ;
 
 
     wxBitmap                 *m_ptemp_icon;
